requirements.txt
flask==2.0.1
flask-login==0.5.0
flask-sqlalchemy==2.5.1
google-cloud-ndb
Authlib
flask_session